Collection of bug fixes (#42)
* fix: force transplants should rise exception flag

* fix: loadstore address and data

* fix: wrong 32bit instruction truncation

* log: remove excess AXI transactions

* fix: page table index function

* fix: the size of the test input operands should be 32bit

* log: disable logging for ipc transactions

* assert: monitor access from cache and MMU

+++ src/test/scala/util/UnitTests.scala
@@ -19,9 +19,9 @@ class TestMultAddBlackbox extends AnyFlatSpec with ChiselScalatestTester {
 
     test(new MACC(32)).withAnnotations(annos) { macc =>
       for( i <- 0 until 1000) {
-        macc.io.mult1.poke(r.nextLong().S)
-        macc.io.mult2.poke(r.nextLong().S)
-        macc.io.add.poke(r.nextLong().S)
+        macc.io.mult1.poke(r.nextInt().S)
+        macc.io.mult2.poke(r.nextInt().S)
+        macc.io.add.poke(r.nextInt().S)
         macc.clock.step()
       }
     }
